Make traverser return a new instance for each file
As the traverser is now stateful, even though it won't break things
as the code is robust enough right now, it severely impacts performances.

For example, the collection of namespace was kept which led to a lot of
namespaces at some point forcing the NamespaceCollection to go through
the node parents to find the proper namespace, even though only one
namespace was found in the file.