Add an init command (#114)
Add an init command to generate a configuration file listing all the
configuration keys with their documentations.

While running the add-prefix command — provided the `--no-config` option
is not used, and no config file is found, the user will be prompted
a messager asking whether or not he wishes to generate the init command.

For now the init command provides a limited interactive experience as
little configuration is convenient to configure from the command line
in an easy way. We are also not entirely sure of how those configuration
will evolve before reaching a stable version.