Rework config and command
Reworked the configuration to make it easier to re-use: the command now
provides the globalNamespaceWhitelister ready for use as well as the
list of files with their contents.

As the handler has a lot of IO-bound dependencies, it is actually very
hard to re-use so it makes more sense to integrate it back to the
command.