Add file contents to the Scoper
Continuing some work for humbug/box#31. The Scoper now takes the file
contents as an argument. This allows the scoper to be used with
non-exitent files (e.g. virtual files) or files for which we are
modifying the content like in Box.

Another benefit is that it removes the Scopers dependency to the file
system layer.