Fix Infection tests
Infection was "passing" only because the process was failing hard which indeeds counts as killing a mutant,
but in this specific case this was not for the right reasons.

I investigated and managed to apply the right patches to make it work. The e2e tests also no runs the not
prefixed Infection and compares the result with the prefixed one as they should be identical.