Fix the scoping of JSON files (#207)
The JSON files should be decode with `assoc` set to true. Otherwise encoding it again may change the schema as if `assoc` is set to false, both `[]` and `{}` will be decoded as an empty array so
encoding it again may change an empty object into an empty array which may screw up a JSON schema validation.