Do not prefix native functions and constants (#280)
The issue reported in #278 is not exact or rather it is the expected behaviour. A case where one will not want that behaviour is for a legitimate internal symbol for which there is two solutions:

- Patch Roave\BetterReflection (which is a lot of work and for which #177 is open)
- Patch the `Reflector` class to account for common cases

Since the first solution is, although more viable, a lot of work and more of a long-term plan, I think the second solution for the meantime is the way to go.