Add make command to dump the COMPOSER_ROOT_VERSION
Since PHP-Scoper depends (as a dev dep) on Box which itself depends on PHP-Scoper, the
COMPOSER_ROOT_VERSION environment variable must be set accordingly for Composer.

Until now, the value was hardcoded in a few places and manually updated. It's however tedious to
keep track of it and as a result, I'm introducing this new command. Whilst it still requires a
manual update, updating the value is now done in a simpler way without having to care about where it
is used and no longer require to manually check what the value should be.