Add a 'on-existing-output-dir' configuration key.
This key can have the values 'ask' (default), 'overwrite' and 'abort',
and controls what happens when the output directory already exists.
The '--force' command line option still overrides it.

Also, 'validateOutputDir' in 'AddPrefixCommand' now returns a bool
(false if directory exists and either the new config is 'abort' or
it's 'ask' and the user cancels). When false is returned the scoping
process is aborted and 'execute' returns an error code of 1.