Add test for #403
We now provide a "polyfill" `Iterator` interface.
This one is not used in any implementation but will get added to the
white list and therefore checked for existence upon autoloader
bootstrap.
In #403, this caused a warning - with the fix, the interface is ignored
(for all PHP versions >=5.0 - hence, all versions).