Move the prefix, patchers and whitelist out of the Scoper interface (#555)
At the time the interface was created, PHPScoper was much simpler and at that time I liked to pass all the parameters which were defining the scoping configuration. The few services that existed were completely independent from the configuration, which means in theory one could scope several projects with the same application, only needing to use a different configuration.

A lot changed however and we are now to a point were more services need to be added to provide a better API built on top of the configuration itself. In this regard, the current design is becoming overly limiting/cumbersome and this already partially shows in Box.

For this reason I am removing the prefix, patchers & whitelist from the `Scoper` interface, which results in an interface much closer to what we have in Box (and potentially will lead to it being removed – it is not the goal itself but more of a side-effect of this change).