Add multi-const statement support (#571)
Previously const statements declaring multiple constants could not be supported (i.e. one of the declared constant needed to be exposed, it would fail).

This is now circumvented by transforming the multi-constant declaration statement into multiple one constant declaration statements. It is a bit ugly as it requires an empty if statement to make it work but otherwise works fine.