feat: Add an inspect command
This adds a command akin to `box process` which allows to output the
scopped content of a singular file.

The command takes similar input as the `add-prefix` regarding the
configuration and behaves identically except it will not try to
execute the `init` command.

The scopped content can also be used as raw input by passing `--quiet`.

Closes #771.