feat: Do not swallow exceptions when scoping (#869)
Closes #847.

As reported in #847, the current strategy is not ideal: if a failure occurs, the file is preserved. You can only see that there was a failure by paying attention to the output or having a more verbose output, unless you pass the `--stop-on-failure` option.

I think historically it has been that way because in the early days of PHP-Scoper, a failure was non infrequent and it was very frustrating to not be able to examine the result despite the failure.

The tool is now more robust and the only exception that I can see regularly coming still is invalid PHP code (e.g. when it is a template). As such, this PR proposes to:

- Leave the file unchanged if a PHP-Parser parsing error occurred (it is still marked as failed and appears in the logs as before).
- Fail on any other failure (i.e. proceed as if `--stop-on-failure` was always provided).
- Introduce a new counterpart option `--continue-on-failure`.

The option `--stop-on-failure` has now been deprecated and will be removed a future version as it is now useless.