test: Do not hide the original throwable stack trace when failing to parse a file
When a spec file could not be parsed, a `self::fail()` was used instead
of letting the throwable bubble up. This was done to know which fail
failed to parsed.

The implementation however, is faulty as if an exception is thrown,
there is no stack trace. This is especially annoying when the throwable
may not be thrown due to an invalid spec file but due to the way this
file is parsed (i.e. code within the `PhpScoperSpecTest`).

This PR introduces a new exception which will contain the previously
thrown exception but will also include a more friendly message to easily
know for which file this was caused.