test: Introduce a Meta object to replace the spec associative array (#985)
At the time the specs were written, there were much less configuration settings and I wanted a clear view of the default settings, not having them declared in an less obvious place (e.g. the spec test case).

Now with the advent of readonly classes, promoted properties and named parameters, I think it is time to reconsider this. This will also make it easier to introduce new settings like needed for #981.